chore: rename module trip_purposes to trip_purpose_enum
