+++ ra2ce/analysis/analysis_config_data/enums/trip_purpose_enum.py
